fixing UI
