+++ ui/src/components/SelectImportPage/screens/ConfirmationScreen.tsx
@@ -34,6 +34,7 @@ export const ConfirmationScreen = ({
   importableComponentTypes,
   teamsResult,
   selectProjectTeam,
+  isOnboardingFlow,
 }: Props & SelectedProjectsProps) => {
   return (
     <>
@@ -81,6 +82,7 @@ export const ConfirmationScreen = ({
         importableComponentTypes={importableComponentTypes}
         teamsResult={teamsResult}
         selectProjectTeam={selectProjectTeam}
+        isOnboardingFlow={isOnboardingFlow}
       />
       {projectsImportingData?.errors && projectsImportingData.errors.length !== 0 && (
         <ErrorWrapper>
